Completed the project - Foundations
It is considered that the game never ends until user or computer reaches
a score of 5 points. The game should check that the user introduces a correct word; maybe
the prompts should keep appearing until "cancel" is clicked or a correct option is typed by the user.
Besides that, the game works as intended.